more tidy up and dont reuse query creation logic
