--- packages/server/src/ai/tools/base/ToolSource.ts
@@ -1,66 +1,0 @@
-import type { ExecutableTool } from ".."
-
-type AgentToolSource = Record<string, any>
-
-/**
- * Abstract base class for tool sources
- * Provides a centralized way to handle tool source operations
- */
-export abstract class ToolSource {
-  protected toolSource: AgentToolSource
-
-  constructor(toolSource: AgentToolSource) {
-    this.toolSource = toolSource
-  }
-
-  /**
-   * Get the type identifier for this tool source
-   */
-  abstract getType(): string
-
-  /**
-   * Get all available tools for this tool source
-   * Implementations should handle authentication and configuration
-   */
-  abstract getTools(): ExecutableTool[]
-
-  /**
-   * Get all available tools asynchronously
-   * Override this for tool sources that need async initialization (e.g., database queries)
-   * Default implementation falls back to sync getTools()
-   */
-  async getToolsAsync(): Promise<ExecutableTool[]> {
-    return this.getTools()
-  }
-
-  /**
-   * Get filtered tools based on disabled tools list
-   */
-  getEnabledTools(): ExecutableTool[] {
-    return this.getTools()
-  }
-
-  /**
-   * Get filtered tools asynchronously based on disabled tools list
-   */
-  async getEnabledToolsAsync(): Promise<ExecutableTool[]> {
-    return this.getToolsAsync()
-  }
-
-  /**
-   * Get guidelines for using this tool source
-   */
-  getGuidelines(): string | undefined {
-    return this.toolSource.auth?.guidelines
-  }
-
-  /**
-   * Validate that the tool source has the required configuration
-   */
-  abstract validate(): boolean
-
-  /**
-   * Get a human-readable name for this tool source
-   */
-  abstract getName(): string
-}
