--- packages/server/src/ai/tools/base/ToolSourceRegistry.ts
@@ -1,27 +1,0 @@
-import { ToolSource } from "./ToolSource"
-import { BudibaseToolSource } from "./BudibaseToolSource"
-import { RestQueryToolSource } from "./RestQueryToolSource"
-
-type AgentToolSource = Record<string, any>
-
-type ToolSourceConstructor = new (toolSource: AgentToolSource) => ToolSource
-
-const registry: Record<string, ToolSourceConstructor> = {
-  BUDIBASE: BudibaseToolSource,
-  REST_QUERY: RestQueryToolSource,
-}
-
-/**
- * Create a tool source instance for the given type
- */
-export function createToolSource(
-  toolSource: AgentToolSource
-): ToolSource | null {
-  const ToolSourceClass = registry[toolSource.type]
-  if (!ToolSourceClass) {
-    console.warn(`Unknown tool source type: ${toolSource.type}`)
-    return null
-  }
-
-  return new ToolSourceClass(toolSource)
-}
