--- packages/server/src/ai/tools/base/index.ts
@@ -1,4 +1,0 @@
-export { ToolSource } from "./ToolSource"
-export { BudibaseToolSource } from "./BudibaseToolSource"
-export { RestQueryToolSource } from "./RestQueryToolSource"
-export * from "./ToolSourceRegistry"
